Move applyKotlinJsImplicitDependencyWorkaround task to the build logic
