+++ build-logic/convention/src/main/kotlin/com/skydoves/landscapist/ComposeMultiplatform.kt
@@ -21,9 +21,11 @@ package com.skydoves.landscapist
 import com.android.build.api.dsl.CommonExtension
 import java.io.File
 import org.gradle.api.Project
+import org.gradle.api.Task
 import org.gradle.api.artifacts.VersionCatalogsExtension
 import org.gradle.kotlin.dsl.dependencies
 import org.gradle.kotlin.dsl.getByType
+import org.gradle.kotlin.dsl.invoke
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import org.jetbrains.kotlin.gradle.targets.js.dsl.ExperimentalWasmDsl
 
@@ -104,7 +106,7 @@ internal fun Project.configureComposeMultiplatform(
     }
 
     explicitApi()
-    task("testClasses")
+    applyKotlinJsImplicitDependencyWorkaround()
   }
 
   commonExtension.apply {
@@ -117,5 +119,44 @@ internal fun Project.configureComposeMultiplatform(
         excludes.add("/META-INF/{AL2.0,LGPL2.1}")
       }
     }
+  }
+}
+
+// https://youtrack.jetbrains.com/issue/KT-56025
+fun Project.applyKotlinJsImplicitDependencyWorkaround() {
+  tasks {
+    val configureJs: Task.() -> Unit = {
+      dependsOn(named("jsDevelopmentLibraryCompileSync"))
+      dependsOn(named("jsDevelopmentExecutableCompileSync"))
+      dependsOn(named("jsProductionLibraryCompileSync"))
+      dependsOn(named("jsProductionExecutableCompileSync"))
+      dependsOn(named("jsTestTestDevelopmentExecutableCompileSync"))
+
+      dependsOn(getByPath(":coil3:jsDevelopmentLibraryCompileSync"))
+      dependsOn(getByPath(":coil3:jsDevelopmentExecutableCompileSync"))
+      dependsOn(getByPath(":coil3:jsProductionLibraryCompileSync"))
+      dependsOn(getByPath(":coil3:jsProductionExecutableCompileSync"))
+      dependsOn(getByPath(":coil3:jsTestTestDevelopmentExecutableCompileSync"))
+    }
+    named("jsBrowserProductionWebpack").configure(configureJs)
+    named("jsBrowserProductionLibraryDistribution").configure(configureJs)
+    named("jsNodeProductionLibraryDistribution").configure(configureJs)
+
+    val configureWasmJs: Task.() -> Unit = {
+      dependsOn(named("wasmJsDevelopmentLibraryCompileSync"))
+      dependsOn(named("wasmJsDevelopmentExecutableCompileSync"))
+      dependsOn(named("wasmJsProductionLibraryCompileSync"))
+      dependsOn(named("wasmJsProductionExecutableCompileSync"))
+      dependsOn(named("wasmJsTestTestDevelopmentExecutableCompileSync"))
+
+      dependsOn(getByPath(":coil3:wasmJsDevelopmentLibraryCompileSync"))
+      dependsOn(getByPath(":coil3:wasmJsDevelopmentExecutableCompileSync"))
+      dependsOn(getByPath(":coil3:wasmJsProductionLibraryCompileSync"))
+      dependsOn(getByPath(":coil3:wasmJsProductionExecutableCompileSync"))
+      dependsOn(getByPath(":coil3:wasmJsTestTestDevelopmentExecutableCompileSync"))
+    }
+    named("wasmJsBrowserProductionWebpack").configure(configureWasmJs)
+    named("wasmJsBrowserProductionLibraryDistribution").configure(configureWasmJs)
+    named("wasmJsNodeProductionLibraryDistribution").configure(configureWasmJs)
   }
 }
